+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Added
 ## Changed
 ## Fixed
+- Fixed #22 by adding a fallback `override` decator for Python < 3.12
 
 # v0.12.1
 ## Fixed